backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pgvector
python-jose[cryptography]
bcrypt
passlib
python-dotenv
pydantic[email]
pydantic-settings

langchain
langchain-community
langchain-groq
sentence-transformers
pymupdf
python-docx
groq
httpx
httpx-sse
python-multipart

apscheduler
reportlab

torch
torchvision
easyocr
opencv-python-headless
numpy
pillow